requirements (2).txt
pyTelegramBotAPI == 4.19.2
python-dotenv == 1.0.1
CurrencyConverter == 0.17.26